requirements.txt
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.38.1
pandas>=2.2.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
numpy>=1.26.0
requests>=2.31.0 